Various improvements
Found a way to remove backup files from router through ssh calling a script with the remove command. Also changed how the various files that need to be added the router are added. This was to clean up the code and consolidate everything into one place. Also added a utility script that will allow me to push the remove script to all of the routers already in our Database.